Token in memory; Token refreshes; silly scheduler
Now the token lives in memory, no longer in a file within the container. That makes it required an initial setup (it's no longer possible to generate the token locally and copy to a container image, which is better anyway.

The token now automatically refreshes. The app uses a configured HTTP client that is able to refresh its token whenever necessary.

There's a silly scheduler that I thought I would need. ¯\_(ツ)_/¯